requirements.txt
FastAPI-SQLAlchemy==0.2.1
Mako==1.2.4
MarkupSafe==2.1.2
PyYAML==6.0
SQLAlchemy==2.0.0
alembic==1.9.2
anyio==3.6.2
attrs==22.2.0
boto3==1.26.62
botocore==1.29.62
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
libmagic==1.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
psycopg2==2.9.5
pydantic==1.10.4
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-magic-bin==0.4.14
python-multipart==0.0.5
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
